ApiClientes/pyproject.toml
[tool.poetry]
name = "casejuntossomosmais"
version = "0.1.0"
description = ""
authors = ["matheus <matheusvictor870@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
phonenumbers = "^8.13.47"
httpx = "^0.27.2"
pydantic = {extras = ["email"], version = "^2.9.2"}
pycountry = "^24.6.1"
django = "^5.1.2"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
uvicorn = "^0.32.0"
fakeredis = {extras = ["json"], version = "^2.25.1"}
redis = "^5.1.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
mypy = "^1.12.0"
ruff = "^0.6.9"
pytest-mock = "^3.14.0"
respx = "^0.21.1"
pytest-cov = "^5.0.0"
pytest-django = "^4.9.0"
pytest-xdist = "^3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "ApiClientes.settings"